2/17 9:06 前台代码

+++ src/main/java/com/zx/redcross/serviceimpl/social/SocialServiceImpl.java
@@ -225,7 +225,9 @@ public class SocialServiceImpl implements SocialService{
 			FileUtils.removeFile(topicType.getThumbnailUrl());
 			return false;
 		}
-		FileUtils.removeFile(thumbnailUrl);
+		if(imgUrl!=null) {
+			FileUtils.removeFile(thumbnailUrl);
+		}
 		return true;
 	}
